apps: move all arguments into a module
Previously the quiche application arguments were split across
three locations. This change combines them into a single module
to make it more straigforward to make updates.
