+++ tools/apps/src/bin/quiche-client.rs
@@ -24,40 +24,11 @@
 // NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+use quiche_apps::args::*;
+
 use quiche_apps::common::*;
 
 use quiche_apps::client::*;
-
-const USAGE: &str = "Usage:
-  quiche-client [options] URL...
-  quiche-client -h | --help
-
-Options:
-  --method METHOD          Use the given HTTP request method [default: GET].
-  --body FILE              Send the given file as request body.
-  --max-data BYTES         Connection-wide flow control limit [default: 10000000].
-  --max-stream-data BYTES  Per-stream flow control limit [default: 1000000].
-  --max-streams-bidi STREAMS  Number of allowed concurrent streams [default: 100].
-  --max-streams-uni STREAMS   Number of allowed concurrent streams [default: 100].
-  --idle-timeout TIMEOUT   Idle timeout in milliseconds [default: 30000].
-  --wire-version VERSION   The version number to send to the server [default: babababa].
-  --http-version VERSION   HTTP version to use [default: all].
-  --dgram-proto PROTO      DATAGRAM application protocol to use [default: none].
-  --dgram-count COUNT      Number of DATAGRAMs to send [default: 0].
-  --dgram-data DATA        Data to send for certain types of DATAGRAM application protocol [default: quack].
-  --dump-packets PATH      Dump the incoming packets as files in the given directory.
-  --dump-responses PATH    Dump response payload as files in the given directory.
-  --dump-json              Dump response headers and payload to stdout in JSON format.
-  --max-json-payload BYTES  Per-response payload limit when dumping JSON [default: 10000].
-  --connect-to ADDRESS     Override ther server's address.
-  --no-verify              Don't verify server's certificate.
-  --no-grease              Don't send GREASE.
-  --cc-algorithm NAME      Specify which congestion control algorithm to use [default: cubic].
-  --disable-hystart        Disable HyStart++.
-  -H --header HEADER ...   Add a request header.
-  -n --requests REQUESTS   Send the given number of identical requests [default: 1].
-  -h --help                Show this screen.
-";
 
 fn main() {
     env_logger::builder()
@@ -65,7 +36,7 @@ fn main() {
         .init();
 
     // Parse CLI parameters.
-    let docopt = docopt::Docopt::new(USAGE).unwrap();
+    let docopt = docopt::Docopt::new(CLIENT_USAGE).unwrap();
     let conn_args = CommonArgs::with_docopt(&docopt);
     let args = ClientArgs::with_docopt(&docopt);
