+++ tools/apps/src/bin/quiche-server.rs
@@ -39,38 +39,11 @@ use std::cell::RefCell;
 
 use ring::rand::*;
 
+use quiche_apps::args::*;
+
 use quiche_apps::common::*;
 
 const MAX_DATAGRAM_SIZE: usize = 1350;
-
-const USAGE: &str = "Usage:
-  quiche-server [options]
-  quiche-server -h | --help
-
-Options:
-  --listen <addr>             Listen on the given IP:port [default: 127.0.0.1:4433]
-  --cert <file>               TLS certificate path [default: src/bin/cert.crt]
-  --key <file>                TLS certificate key path [default: src/bin/cert.key]
-  --root <dir>                Root directory [default: src/bin/root/]
-  --index <name>              The file that will be used as index [default: index.html].
-  --name <str>                Name of the server [default: quic.tech]
-  --max-data BYTES            Connection-wide flow control limit [default: 10000000].
-  --max-stream-data BYTES     Per-stream flow control limit [default: 1000000].
-  --max-streams-bidi STREAMS  Number of allowed concurrent streams [default: 100].
-  --max-streams-uni STREAMS   Number of allowed concurrent streams [default: 100].
-  --idle-timeout TIMEOUT   Idle timeout in milliseconds [default: 30000].
-  --dump-packets PATH         Dump the incoming packets as files in the given directory.
-  --early-data                Enables receiving early data.
-  --no-retry                  Disable stateless retry.
-  --no-grease                 Don't send GREASE.
-  --http-version VERSION      HTTP version to use [default: all].
-  --dgram-proto PROTO         DATAGRAM application protocol to use [default: none].
-  --dgram-count COUNT         Number of DATAGRAMs to send [default: 0].
-  --dgram-data DATA           Data to send for certain types of DATAGRAM application protocol [default: brrr].
-  --cc-algorithm NAME         Specify which congestion control algorithm to use [default: cubic].
-  --disable-hystart           Disable HyStart++.
-  -h --help                   Show this screen.
-";
 
 fn main() {
     let mut buf = [0; 65535];
@@ -81,7 +54,7 @@ fn main() {
         .init();
 
     // Parse CLI parameters.
-    let docopt = docopt::Docopt::new(USAGE).unwrap();
+    let docopt = docopt::Docopt::new(SERVER_USAGE).unwrap();
     let conn_args = CommonArgs::with_docopt(&docopt);
     let args = ServerArgs::with_docopt(&docopt);
 
@@ -575,39 +548,4 @@ fn validate_token<'a>(
     let token = &token[addr.len()..];
 
     Some(&token[..])
-}
-
-// Application-specific arguments that compliment the `CommonArgs`.
-struct ServerArgs {
-    listen: String,
-    no_retry: bool,
-    root: String,
-    index: String,
-    cert: String,
-    key: String,
-    early_data: bool,
-}
-
-impl Args for ServerArgs {
-    fn with_docopt(docopt: &docopt::Docopt) -> Self {
-        let args = docopt.parse().unwrap_or_else(|e| e.exit());
-
-        let listen = args.get_str("--listen").to_string();
-        let no_retry = args.get_bool("--no-retry");
-        let early_data = args.get_bool("--early-data");
-        let root = args.get_str("--root").to_string();
-        let index = args.get_str("--index").to_string();
-        let cert = args.get_str("--cert").to_string();
-        let key = args.get_str("--key").to_string();
-
-        ServerArgs {
-            listen,
-            no_retry,
-            root,
-            index,
-            cert,
-            key,
-            early_data,
-        }
-    }
 }
