+++ tools/apps/src/common.rs
@@ -70,148 +70,6 @@ pub mod alpns {
     }
 }
 
-pub trait Args {
-    fn with_docopt(docopt: &docopt::Docopt) -> Self;
-}
-
-/// Contains commons arguments for creating a quiche QUIC connection.
-pub struct CommonArgs {
-    pub alpns: Vec<u8>,
-    pub max_data: u64,
-    pub max_stream_data: u64,
-    pub max_streams_bidi: u64,
-    pub max_streams_uni: u64,
-    pub idle_timeout: u64,
-    pub dump_packet_path: Option<String>,
-    pub no_grease: bool,
-    pub cc_algorithm: String,
-    pub disable_hystart: bool,
-    pub dgrams_enabled: bool,
-    pub dgram_count: u64,
-    pub dgram_data: String,
-}
-
-/// Creates a new `CommonArgs` structure using the provided [`Docopt`].
-///
-/// The `Docopt` usage String needs to include the following:
-///
-/// --http-version VERSION      HTTP version to use.
-/// --max-data BYTES            Connection-wide flow control limit.
-/// --max-stream-data BYTES     Per-stream flow control limit.
-/// --max-streams-bidi STREAMS  Number of allowed concurrent streams.
-/// --max-streams-uni STREAMS   Number of allowed concurrent streams.
-/// --dump-packets PATH         Dump the incoming packets in PATH.
-/// --no-grease                 Don't send GREASE.
-/// --cc-algorithm NAME         Set a congestion control algorithm.
-/// --disable-hystart           Disable HyStart++.
-/// --dgram-proto PROTO         DATAGRAM application protocol.
-/// --dgram-count COUNT         Number of DATAGRAMs to send.
-///  --dgram-data DATA          DATAGRAM data to send.
-///
-/// [`Docopt`]: https://docs.rs/docopt/1.1.0/docopt/
-impl Args for CommonArgs {
-    fn with_docopt(docopt: &docopt::Docopt) -> Self {
-        let args = docopt.parse().unwrap_or_else(|e| e.exit());
-
-        let http_version = args.get_str("--http-version");
-        let dgram_proto = args.get_str("--dgram-proto");
-        let (alpns, dgrams_enabled) = match (http_version, dgram_proto) {
-            ("HTTP/0.9", "none") =>
-                (alpns::length_prefixed(&alpns::HTTP_09), false),
-
-            ("HTTP/0.9", _) =>
-                panic!("Unsupported HTTP version and DATAGRAM protocol."),
-
-            ("HTTP/3", "none") => (alpns::length_prefixed(&alpns::HTTP_3), false),
-
-            ("HTTP/3", "oneway") =>
-                (alpns::length_prefixed(&alpns::HTTP_3), true),
-
-            ("all", "none") => (
-                [
-                    alpns::length_prefixed(&alpns::HTTP_3),
-                    alpns::length_prefixed(&alpns::HTTP_09),
-                ]
-                .concat(),
-                false,
-            ),
-
-            // SiDuck is it's own application protocol.
-            (_, "siduck") => (alpns::length_prefixed(&alpns::SIDUCK), true),
-
-            (..) => panic!("Unsupported HTTP version and DATAGRAM protocol."),
-        };
-
-        let dgram_count = args.get_str("--dgram-count");
-        let dgram_count = u64::from_str_radix(dgram_count, 10).unwrap();
-
-        let dgram_data = args.get_str("--dgram-data").to_string();
-
-        let max_data = args.get_str("--max-data");
-        let max_data = u64::from_str_radix(max_data, 10).unwrap();
-
-        let max_stream_data = args.get_str("--max-stream-data");
-        let max_stream_data = u64::from_str_radix(max_stream_data, 10).unwrap();
-
-        let max_streams_bidi = args.get_str("--max-streams-bidi");
-        let max_streams_bidi = u64::from_str_radix(max_streams_bidi, 10).unwrap();
-
-        let max_streams_uni = args.get_str("--max-streams-uni");
-        let max_streams_uni = u64::from_str_radix(max_streams_uni, 10).unwrap();
-
-        let idle_timeout = args.get_str("--idle-timeout");
-        let idle_timeout = u64::from_str_radix(idle_timeout, 10).unwrap();
-
-        let dump_packet_path = if args.get_str("--dump-packets") != "" {
-            Some(args.get_str("--dump-packets").to_string())
-        } else {
-            None
-        };
-
-        let no_grease = args.get_bool("--no-grease");
-
-        let cc_algorithm = args.get_str("--cc-algorithm");
-
-        let disable_hystart = args.get_bool("--disable-hystart");
-
-        CommonArgs {
-            alpns,
-            max_data,
-            max_stream_data,
-            max_streams_bidi,
-            max_streams_uni,
-            idle_timeout,
-            dump_packet_path,
-            no_grease,
-            cc_algorithm: cc_algorithm.to_string(),
-            disable_hystart,
-            dgrams_enabled,
-            dgram_count,
-            dgram_data,
-        }
-    }
-}
-
-impl Default for CommonArgs {
-    fn default() -> Self {
-        CommonArgs {
-            alpns: alpns::length_prefixed(&alpns::HTTP_3),
-            max_data: 10000000,
-            max_stream_data: 1000000,
-            max_streams_bidi: 100,
-            max_streams_uni: 100,
-            idle_timeout: 30000,
-            dump_packet_path: None,
-            no_grease: false,
-            cc_algorithm: "cubic".to_string(),
-            disable_hystart: false,
-            dgrams_enabled: false,
-            dgram_count: 0,
-            dgram_data: "quack".to_string(),
-        }
-    }
-}
-
 pub struct PartialRequest {
     pub req: Vec<u8>,
 }
