+++ tools/apps/src/lib.rs
@@ -27,5 +27,6 @@
 #[macro_use]
 extern crate log;
 
+pub mod args;
 pub mod client;
 pub mod common;
